+++ v2/src/dar_backup/__about__.py
@@ -1,4 +1,4 @@
-__version__ = "1.1.10.dev21"
+__version__ = "1.1.10.dev22"
 
 __author__ = "Per Jensen"